+++ src/Farmer/Builders/Builders.Vm.fs
@@ -87,7 +87,7 @@ type VmConfig =
                     StorageAccount = this.DiagnosticsStorageAccount |> Option.map (fun r -> r.resourceId(this).Name)
                     NetworkInterfaceName = this.NicName.Name
                     Size = this.Size
-                    Priority = this.Priority |> Option.defaultValue Regular
+                    Priority = this.Priority
                     Credentials =
                         match this.Username with
                         | Some username ->
@@ -442,6 +442,11 @@ type VirtualMachineBuilder() =
     member this.LinkToVNet(state: VmConfig, vnet: VirtualNetworkConfig) = this.LinkToVNet(state, vnet.Name)
 
     [<CustomOperation "link_to_unmanaged_vnet">]
+    member _.LinkToUnmanagedVNet(state: VmConfig, id: ResourceId) =
+        { state with
+            VNet = LinkedResource(Unmanaged(id))
+        }
+
     member _.LinkToUnmanagedVNet(state: VmConfig, name: ResourceName) =
         { state with
             VNet = LinkedResource(Unmanaged(virtualNetworks.resourceId name))