+++ src/Farmer/Common.fs
@@ -151,6 +151,22 @@ type EnvVar =
 module Mb =
     let toBytes (mb: int<Mb>) = int64 mb * 1024L * 1024L
 
+module Route =
+    type HopType =
+        | VirtualAppliance of System.Net.IPAddress option
+        | Internet
+        | Nothing
+        | VirtualNetworkGateway
+        | VnetLocal
+
+        member x.ArmValue =
+            match x with
+            | VirtualAppliance _ -> "VirtualAppliance"
+            | Internet -> "Internet"
+            | Nothing -> "None"
+            | VirtualNetworkGateway -> "VirtualNetworkGateway"
+            | VnetLocal -> "VnetLocal"
+
 module Vm =
     type VMSize =
         | Basic_A0
@@ -1818,9 +1834,13 @@ module Identity =
     /// Represents a User Assigned Identity, and the ability to create a Principal Id from it.
     type UserAssignedIdentity =
         | UserAssignedIdentity of ResourceId
+        | LinkedUserAssignedIdentity of ResourceId
 
         member private this.CreateExpression field =
-            let (UserAssignedIdentity resourceId) = this
+            let resourceId =
+                match this with
+                | UserAssignedIdentity rid -> rid
+                | LinkedUserAssignedIdentity rid -> rid
 
             ArmExpression
                 .create($"reference({resourceId.ArmExpression.Value}).%s{field}")
@@ -1831,7 +1851,8 @@ module Identity =
 
         member this.ResourceId =
             match this with
-            | UserAssignedIdentity r -> r
+            | UserAssignedIdentity rid -> rid
+            | LinkedUserAssignedIdentity rid -> rid
 
     type SystemIdentity =
         | SystemIdentity of ResourceId
@@ -1857,7 +1878,12 @@ module Identity =
             UserAssigned: UserAssignedIdentity list
         }
 
-        member this.Dependencies = this.UserAssigned |> List.map (fun u -> u.ResourceId)
+        member this.Dependencies =
+            this.UserAssigned
+            |> List.choose (fun identity ->
+                match identity with
+                | UserAssignedIdentity rid -> Some rid
+                | LinkedUserAssignedIdentity _ -> None)
 
         static member Empty =
             {
@@ -1875,6 +1901,8 @@ module Identity =
             { managedIdentity with
                 UserAssigned = userAssignedIdentity :: managedIdentity.UserAssigned
             }
+
+open Identity
 
 module Containers =
     type DockerImage =
@@ -1911,6 +1939,7 @@ type ImageRegistryCredential =
         Server: string
         Username: string
         Password: SecureParameter
+        Identity: ManagedIdentity
     }
 
 [<RequireQualifiedAccess>]
@@ -1919,6 +1948,8 @@ type ImageRegistryAuthentication =
     | Credential of ImageRegistryCredential
     /// Credentials for the container registry will be listed by ARM expression.
     | ListCredentials of ResourceId
+    /// Credentials for the container registry are included with the identity as a template parameter.
+    | ManagedIdentityCredential of ImageRegistryCredential
 
 [<RequireQualifiedAccess>]
 type LogAnalyticsWorkspace =